+++ src/core/io/lottie/lottie_exporter.hpp
@@ -411,6 +411,51 @@ public:
         return QCborArray{cb};
     }
 
+    std::vector<std::unique_ptr<model::KeyframeBase>> split_keyframes(model::AnimatableBase* prop)
+    {
+        std::vector<std::unique_ptr<model::KeyframeBase>> split_kfs;
+        std::unique_ptr<model::KeyframeBase> previous = prop->keyframe(0)->clone();
+
+        for ( int i = 1, e = prop->keyframe_count(); i < e; i++ )
+        {
+            if ( previous->transition().hold() )
+            {
+                split_kfs.push_back(std::move(previous));
+                previous = prop->keyframe(1)->clone();
+                continue;
+            }
+
+            std::array<qreal, 2> raw_splits;
+            std::tie(raw_splits[0], raw_splits[1]) = previous->transition().bezier().extrema(1);
+            qDebug() << previous->transition().bezier().points()[1] << previous->transition().bezier().points()[2];
+
+            std::vector<qreal> splits;
+            for ( qreal t : raw_splits )
+            {
+                QPointF p = previous->transition().bezier().solve(t);
+                if ( p.y() < 0 || p.y() > 1 )
+                    splits.push_back(t);
+            }
+
+            if ( splits.size() == 0 )
+            {
+                split_kfs.push_back(std::move(previous));
+                previous = prop->keyframe(i)->clone();
+            }
+            else
+            {
+                auto next = prop->keyframe(1);
+                auto next_segment = previous->split(next, splits);
+                previous = std::move(next_segment.back());
+                split_kfs.insert(split_kfs.end(), std::make_move_iterator(next_segment.begin()), std::make_move_iterator(next_segment.end() - 1));
+            }
+        }
+
+        split_kfs.push_back(std::move(previous));
+
+        return split_kfs;
+    }
+
     QCborMap convert_animated(
         model::AnimatableBase* prop,
         const TransformFunc& transform_values
@@ -422,11 +467,13 @@ public:
         if ( prop->keyframe_count() > 1 )
         {
             jobj["a"_l] = 1;
+            std::vector<std::unique_ptr<model::KeyframeBase>> split_kfs = split_keyframes(prop);
+
             QCborArray keyframes;
             QCborMap jkf;
-            for ( int i = 0, e = prop->keyframe_count(); i < e; i++ )
+            for ( int i = 0, e = split_kfs.size(); i < e; i++ )
             {
-                auto kf = prop->keyframe(i);
+                auto kf = split_kfs[i].get();
                 QVariant v = transform_values.to_lottie(kf->value(), kf->time());
                 QCborValue kf_value = keyframe_value_from_variant(v);