+++ src/core/model/animation/animatable_path.cpp
@@ -15,7 +15,11 @@ void glaxnimate::model::detail::AnimatedPropertyBezier::set_closed(bool closed)
 {
     value_.set_closed(closed);
     for ( auto& keyframe : keyframes_ )
-        keyframe->value_.set_closed(closed);
+    {
+        auto v = keyframe->get();
+        v.set_closed(closed);
+        keyframe->set(v);
+    }
     value_changed();
     emitter(object(), value_);
 }
@@ -61,7 +65,7 @@ void glaxnimate::model::detail::AnimatedPropertyBezier::remove_points(const std:
     bool set = true;
     for ( const auto& kf : keyframes_ )
     {
-        auto bez = kf->value_.removed_points(indices);
+        auto bez = kf->get().removed_points(indices);
         if ( !mismatched_ && kf->time() == time() )
             set = false;
         object()->push_command(new command::SetKeyframe(this, kf->time(), QVariant::fromValue(bez), true));
@@ -143,7 +147,7 @@ void glaxnimate::model::detail::AnimatedPropertyBezier::extend(const math::bezie
         if ( !mismatched_ && kf->time() == time() )
             set = false;
         object()->push_command(
-            new command::SetKeyframe(this, kf->time(), extend_impl(kf->value_, target, at_end), true)
+            new command::SetKeyframe(this, kf->time(), extend_impl(kf->get(), target, at_end), true)
         );
     }