+++ src/gui/widgets/timeline/keyframe_editor_widget.cpp
@@ -24,6 +24,7 @@ void set_icon(QComboBox* box, int i, const char* ba)
         case model::KeyframeTransition::Linear: which = "linear"; break;
         case model::KeyframeTransition::Ease: which = "ease"; break;
         case model::KeyframeTransition::Fast: which = "fast"; break;
+        case model::KeyframeTransition::Overshoot: which = "overshoot"; break;
         case model::KeyframeTransition::Custom: which = "custom"; break;
     }
     box->setItemIcon(i, QIcon(app::Application::instance()->data_file(icon_name.arg(ba).arg(which))));